Fix a memory leak in circular lists
When a circular list is cleared, its 'count' value is set to 0. The
pointers to elements that were present in the list are still in the
'values', which is expected.

But when adding a new element, the code should always check if there
already is an element in a cell, even if the list appears to be not full
(i.e. list->count == list->size is false), otherwise the elements left
there after clearing the list would be leaked.