Fix UNIX to WINDOWS cross-compilation for LuaG library v2
Before the recent changes to the cross-compilation system, the windows
version of LuaG Console was first compiled as a shared library and then
an executable wrapper would call the main function. The LuaG libraries
were then linked against that shared library.

Now, an Import Library of LuaG Console is generated, and LuaG libraries
can link against it.